[REFACTOR] #41 - data class 폴더링 변경
