--- app/src/main/java/org/cardna/data/remote/api/Temp.kt
@@ -1,4 +1,0 @@
-package org.cardna.data.remote.api
-
-class Temp {
-}
