+++ app/src/main/java/org/cardna/ui/maincard/adapter/RepresentBottomSheetCardMeAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.core.view.isVisible
 import androidx.recyclerview.widget.RecyclerView
-import org.cardna.data.remote.api.RepresentCardData
+import org.cardna.data.remote.model.representcardedit.RepresentCardData
 import org.cardna.databinding.ItemRepresentCardMeBinding
 
 class RepresentBottomSheetCardMeAdapter :
