+++ app/src/main/res/layout/item_main_card_view.xml
@@ -15,10 +15,11 @@
 
         <ImageView
             android:id="@+id/iv_main_card_image"
-            android:layout_width="match_parent"
+            android:layout_width="0dp"
             android:layout_height="0dp"
             android:layout_marginHorizontal="8dp"
             android:layout_marginTop="8dp"
+
             android:scaleType="fitXY"
             android:layout_marginBottom="4dp"
             android:src="@drawable/dummy_img_test"
